Switch to MailHog and move behind reverse proxy
Unfortunately mailcatcher is not well-suited to being behind a reverse
proxy so I've switched to use MailHog instead.

The MailHog image is much simpler to proxy and gives the same kind of
result in a very lightweight image.

The benefit of using a reverse proxy is that we do not tie directly to
ports, but instead have a single port for the whole install making
parallel runs more achievable.

Fixed #24